Update to 1.16.4

+++ src/main/java/com/mrbysco/insane/capability/SanityCapability.java
@@ -7,7 +7,7 @@ import java.math.RoundingMode;
 
 public class SanityCapability implements ISanity {
     public static double SANITY_MIN = 0.0D;
-    public static double SANITY_MAX = InsaneConfig.COMMON.maxSanity.get().doubleValue();
+    public static double SANITY_MAX = InsaneConfig.COMMON.maxSanity.get();
     private double sanity;
     private boolean dirty;
